refactor: extract dendrite helpers to reduce duplication
